Tracking Problem
This iteration of the code looks at tracking position and attitude for the thrust vectoring quadrotor. The controller used for this problem is feedback linearization.

This problem looks at testing the code without the control allocation. The resulting desired forces and moments from the controller are directly fed into the dynamics of the system.