Tracking Problem With Control Allocation
Changes:

When Calculating T  = k*w^2, it was changes to T = k *(w*(pi/30))^2 in order for the magnitudes of the desired  and actual virtual control inputs.

In terms of the sign, in the function virtualControl: The MPhi and MThe signs are manually flipped. This looks to be a temporary solution, but needs to be further analyzed.

A new folder FBL- with CA is created to indicate the progress made using the feedback linearization controller and the implementation of the Control Allocation.